Rename teams.ts to teamsData.ts
